requirements.txt
jupyter
pandas
python-dotenv
openai
openai
pdfplumber
pandas
openpyxl
